WIP: Initial orchestrator implementation.
This is the WIP of the orchestration component.

The basic model is as follows:

All IPs generate a unique leadership challenge in the orchestrator.

If the orchestrator wins the challenge, the workload deployment proceeds.
When the orchestrator loses leadership, it releases the IP (if Calico thinks
this host has it) and possibly kills any workload in docker which has it.